requirements.txt
openai
anthropic
requests
datasets
scikit-learn
matplotlib
python-dotenv